Split sysfs actor by concern
Reorganize the sysfs actor into separate files for the public runtime
API, worker setup and routing, worker execution, and config building.

Introduce a small internal workerSet so worker routes and lifecycle
tracking stay grouped without changing runtime behavior.
